preset images and better modal class window
You can now add ILSVRC2012 images in the image folder and a ground truth file will be used to automatically set the image class code. If a model classifies a preset image incorrectly, an error message will be shown. The alert message to confirm the image prediction was substituted with a bootstrap modal window.